completed quizfunctions and added authentication to them

+++ Project/QuizManager/src/controllers/quiz.ts
@@ -29,13 +29,20 @@ const createQuiz=async (req:Request, res:Response, next:NextFunction) =>{
 const getQuiz=async (req:Request, res:Response, next:NextFunction) =>{
     try {
         const quizId=req.params.quizId;
-        const quiz=await Quiz.findById(quizId,{name:1, questions_list:1, answers:1});
+        const quiz=await Quiz.findById(quizId,{name:1, questions_list:1, answers:1, created_by:1});
 
         if(!quiz){
             const err=new ProjectError("Quiz not found");
             err.statusCode=404;
             throw err;
         }
+
+        if(req.userId!==quiz.created_by.toString()){
+            const err=new ProjectError("You are not Authorized");
+            err.statusCode=403;
+            throw err;
+        }
+
         const resp:ReturnResponse={status:"success", message:"Quiz retrieved successfully", data:{quiz}};
         res.status(201).send(resp);
 
@@ -50,6 +57,11 @@ const updateQuiz=async (req:Request, res:Response, next:NextFunction) =>{    try
         if(!quiz){
             const err=new ProjectError("Quiz not found");
             err.statusCode=404;
+            throw err;
+        }
+        if(req.userId!==quiz.created_by.toString()){
+            const err=new ProjectError("You are not Authorized");
+            err.statusCode=403;
             throw err;
         }
         quiz.name=req.body.name;
@@ -67,6 +79,18 @@ const updateQuiz=async (req:Request, res:Response, next:NextFunction) =>{    try
 const deleteQuiz=async (req:Request, res:Response, next:NextFunction) =>{
     try {
         const quizId=req.params.quizId;
+        const quiz=await Quiz.findById(quizId);
+        if(!quiz){
+            const err=new ProjectError("Quiz not found");
+            err.statusCode=404;
+            throw err;
+        }
+        if(req.userId!==quiz.created_by.toString()){
+            const err=new ProjectError("You are not Authorized");
+            err.statusCode=403;
+            throw err;
+        }
+
         await Quiz.deleteOne({_id:quizId});
         const resp:ReturnResponse={status:"success", message:"Quiz deleted successfully", data:{}};
         res.status(200).send(resp);
@@ -77,7 +101,23 @@ const deleteQuiz=async (req:Request, res:Response, next:NextFunction) =>{
 
 const publishQuiz=async (req:Request, res:Response, next:NextFunction) =>{
     try {
-        const quizId=req.params.quizId;
+        const quizId=req.body.quizId;
+        const quiz=await Quiz.findById(quizId);
+        if(!quiz){
+            const err=new ProjectError("Quiz not found");
+            err.statusCode=404;
+            throw err;
+    }
+    if(req.userId!==quiz.created_by.toString()){
+        const err=new ProjectError("You are not Authorized");
+        err.statusCode=403;
+        throw err;
+    }
+
+    quiz.is_published=true;
+    await quiz.save();
+    const resp:ReturnResponse={status:"success", message:"Quiz published", data:{}};
+    res.status(200).send(resp);
     }
      catch (error) {
         next(error);
